Migrate to postgres

+++ olmonoko-backend/migrations/20240705153404_ics_sources_hash.up.sql
@@ -1,3 +1,3 @@
 ALTER TABLE ics_sources ADD COLUMN file_hash TEXT;
 ALTER TABLE ics_sources ADD COLUMN object_hash TEXT;
-ALTER TABLE ics_sources ADD COLUMN updated_at INTEGER;
+ALTER TABLE ics_sources ADD COLUMN updated_at BIGINT;
